+++ src/Zeta/Validators/FieldContextlessValidator.cs
@@ -1,14 +1,14 @@
 namespace Zeta.Validators;
 
-internal sealed class FieldContextlessValidator<TInstance, TProperty> : IFieldContextlessValidator<TInstance> 
+internal sealed class FieldContextlessValidator<TInstance, TProperty> : IFieldContextlessValidator<TInstance>
 {
     private static readonly IReadOnlyList<ValidationError> EmptyErrors = [];
 
     private readonly string _name;
-    private readonly Func<TInstance, TProperty?> _getter;
+    private readonly Func<TInstance, TProperty> _getter;
     private readonly ISchema<TProperty> _schema;
 
-    public FieldContextlessValidator(string name, Func<TInstance, TProperty?> getter, ISchema<TProperty> schema)
+    public FieldContextlessValidator(string name, Func<TInstance, TProperty> getter, ISchema<TProperty> schema)
     {
         _name = name;
         if (!string.IsNullOrEmpty(_name) && char.IsUpper(_name[0]))